requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
onnx==1.15.0
onnx-tf==1.10.0
onnx2keras==0.0.24
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==4.24.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch2keras==0.2.4
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.22.1
termcolor==2.3.0
torch==2.1.0
torchvision==0.16.0
traitlets==5.13.0
typeguard==2.13.3
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.14.1
